requirements.txt
absl-py==0.7.1
alabaster==0.7.10
anaconda-navigator==1.8.7
anaconda-project==0.8.2
anykeystore==0.2
apex==0.1
asn1crypto==0.24.0
astor==0.8.0
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
backtrace==0.2.1
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
Bottleneck==1.2.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.7
conda==4.8.3
conda-build==3.10.5
conda-package-handling==1.7.0
conda-verify==2.0.0
contextlib2==0.5.5
cryptacular==1.5.5
cryptography==2.9.2
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
d2lzh==0.8.11
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
defusedxml==0.6.0
distributed==1.21.8
docutils==0.14
easydict==1.9
entrypoints==0.2.3
et-xmlfile==1.0.1
facenet-pytorch==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
future==0.18.0
gast==0.2.2
gevent==1.3.0
glob2==0.6
gmpy2==2.0.8
google-pasta==0.1.7
graphviz==0.8.4
greenlet==0.4.13
grpcio==1.21.1
h5py==2.8.0rc1
heapdict==1.0.0
html5lib==1.0.1
hupper==1.9.1
idna==2.6
imageio==2.3.0
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
json5==0.8.5
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.0.2
jupyterlab-launcher==0.10.5
jupyterlab-server==1.0.0
jupyterthemes==0.20.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.3.1
lesscpy==0.13.0
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.6.1
msgpack-python==0.5.6
multipledispatch==0.5.0
mxnet-cu100==1.4.1
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==6.0.0
numba==0.38.0
numexpr==2.6.5
numpy==1.18.4
numpydoc==0.8.0
oauthlib==3.1.0
odo==0.5.1
olefile==0.45.1
opencv-python==4.1.0.25
openpyxl==2.5.3
opt-einsum==3.2.1
packaging==17.1
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
PasteDeploy==2.0.1
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pbkdf2==1.3
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.4.1
pkginfo==1.4.2
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.6.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.8.0
psutil==5.4.5
ptyprocess==0.5.2
py==1.5.3
pycocotools==2.0.0
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.5
pydot==1.4.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyramid==1.10.4
pyramid-mailer==0.15.1
pyrsistent==0.15.3
PySnooper==0.2.2
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
repoze.sendmail==4.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
shy==0.2.3
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tb-nightly==1.14.0a20190603
tblib==1.3.2
tensorboard==2.0.0
tensorboardX==1.8
tensorflow==2.0.0b0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.0
terminator==0.0.0
testpath==0.3.1
tf-estimator-nightly==1.14.0.dev2019060501
toolz==0.9.0
torch==1.1.0
TorchSnooper==0.4.1
torchvision==0.3.0
tornado==5.0.2
tqdm==4.46.0
traitlets==4.3.2
transaction==2.4.0
translationstring==1.3
typing==3.6.4
ujson==1.35
unicodecsv==0.14.1
urllib3==1.22
urlparser==0.1.2
velruse==1.1.1
venusian==3.0.0
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.11.2
WTForms==2.2.1
wtforms-recaptcha==0.3.2
xgboost==0.90
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
zict==0.1.3
zope.deprecation==4.4.0
zope.interface==4.7.1
zope.sqlalchemy==1.2
